A lot of changes (read desc)
I have made a main menu where you can choose to play loval multiplayer, and choose the amount of players.

I have also remade the cards for the third time, now using a quad parent and two sprites for back and face. Which I had to write extra code for, and change existing.

The game also works better now, names and score of the players are displayed, you can draw cards from the deck (and if empty the turn simply ends), and you can get points for getting four cards of the same number, which then removes them from play.

And firebase stuff